# openapi.yaml
openapi: 3.1.0
info:
  title: IFRS9 API
  version: 1.0.0
paths:
...
  /admin/requests:
    get:
      tags:
      - admin
      summary: Get Access Requests
      description: Get all access requests
      operationId: list_all_access_requests
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AccessRequestResponse'
                type: array
                title: Response List All Access Requests
      security:
      - OAuth2PasswordBearer: []
...
  /admin/requests/{request_id}:
    put:
      tags:
      - admin
      summary: Update Access Request
      description: Update access requests
      operationId: update_specific_access_request
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '404':
          description: Request not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Delete Access Request
      description: Delete access requests
      operationId: delete_access_request_admin_requests__request_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
          title: Request Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Request not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /admin/users:
    get:
      tags:
      - admin
      summary: Get Users
      description: Get all users in db
      operationId: get_users_admin_users_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserResponse'
                type: array
                title: Response Get Users Admin Users Get
      security:
      - OAuth2PasswordBearer: []
    post:
      tags:
      - admin
      summary: Create User
      description: create a user in db
      operationId: create_user_admin_users_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /admin/users/export:
    get:
      tags:
      - admin
      summary: Export Users Csv
      description: Get all users and export them as csv
      operationId: export_users_csv_admin_users_export_get
      responses:
        '200':
          description: Successful Response
      security:
      - OAuth2PasswordBearer: []
  /admin/feedback:
    get:
      tags:
      - admin
      summary: Admin Get All Feedback
      description: Get all feedback entries
      operationId: admin_get_all_feedback_admin_feedback_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
                type: array
                title: Response Admin Get All Feedback Admin Feedback Get
      security:
      - OAuth2PasswordBearer: []
  /admin/help:
    get:
      tags:
      - admin
      summary: Admin Get All Help
      description: Get all help entries
      operationId: admin_get_all_help_admin_help_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/HelpResponse'
                type: array
                title: Response Admin Get All Help Admin Help Get
      security:
      - OAuth2PasswordBearer: []
  /admin/help/{help_id}:
    get:
      tags:
      - admin
      summary: Admin Get Help
      description: Get help entry by ID
      operationId: admin_get_help_admin_help__help_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: help_id
        in: path
        required: true
        schema:
          type: integer
          title: Help Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelpResponse'
        '404':
          description: Help entry not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Admin Delete Help
      description: Delete a specific help entry by ID
      operationId: admin_delete_help_admin_help__help_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: help_id
        in: path
        required: true
        schema:
          type: integer
          title: Help Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Help not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/help/{help_id}/status:
    put:
      tags:
      - admin
      summary: Update Help Status
      description: Get help status by ID
      operationId: update_help_status_admin_help__help_id__status_put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: help_id
        in: path
        required: true
        schema:
          type: integer
          title: Help Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HelpStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelpResponse'
        '404':
          description: Help not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/users/{user_id}:
    get:
      tags:
      - admin
      summary: Get User
      description: Get a specific user by ID
      operationId: get_user_admin_users__user_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - admin
      summary: Update User
      operationId: update_user_admin_users__user_id__put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Delete User
      description: Delete a user by ID
      operationId: delete_user_admin_users__user_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
          title: User Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: User not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/feedback/{feedback_id}:
    get:
      tags:
      - admin
      summary: Admin Get Feedback
      description: Get a specific feedback by ID
      operationId: admin_get_feedback_admin_feedback__feedback_id__get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '404':
          description: Feedback not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - admin
      summary: Admin Delete Feedback
      description: Delete a feedback by ID
      operationId: admin_delete_feedback_admin_feedback__feedback_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Feedback not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/feedback/{feedback_id}/status:
    put:
      tags:
      - admin
      summary: Update Feedback Status
      description: Update a feedback status
      operationId: update_feedback_status_admin_feedback__feedback_id__status_put
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: feedback_id
        in: path
        required: true
        schema:
          type: integer
          title: Feedback Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackStatusUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '404':
          description: Feedback not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AccessRequestResponse:
      properties:
        id:
          type: integer
          title: Id
        email:
          type: string
          format: email
          title: Email
        admin_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Admin Email
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        is_email_verified:
          type: boolean
          title: Is Email Verified
      type: object
      required:
      - id
      - email
      - status
      - created_at
      - is_email_verified
      title: AccessRequestResponse
...
    AccessRequestUpdate:
      properties:
        status:
          $ref: '#/components/schemas/RequestStatus'
        role:
          anyOf:
          - $ref: '#/components/schemas/UserRole'
          - type: 'null'
      type: object
      required:
      - status
      title: AccessRequestUpdate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HelpStatusUpdate:
      properties:
        status:
          $ref: '#/components/schemas/HelpStatusEnum'
      type: object
      required:
      - status
      title: HelpStatusUpdate
      description: Schema for updating the status of a help request (admin only)
...
    UserCreate:
      properties:
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        email:
          type: string
          format: email
          title: Email
        recovery_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Recovery Email
        role:
          $ref: '#/components/schemas/UserRole'
          default: user
        is_active:
          type: boolean
          title: Is Active
          default: true
        portfolio_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Portfolio Id
      type: object
      required:
      - email
      title: UserCreate
...
    UserResponse:
      properties:
        id:
          type: integer
          title: Id
        first_name:
          anyOf:
          - type: string
          - type: 'null'
          title: First Name
        last_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Name
        email:
          type: string
          format: email
          title: Email
        recovery_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Recovery Email
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
          title: Is Active
        created_at:
          type: string
          format: date-time
          title: Created At
        last_login:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Login
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - id
      - email
      - role
      - is_active
      - created_at
      title: UserResponse
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token